+++ main.go
@@ -7,9 +7,7 @@ import (
 	"log"
 	"net/http"
 	"net/http/httputil"
-	"net/url"
 	"os"
-	"strconv"
 
 	"github.com/kaancfidan/bouncer/services"
 )
@@ -20,13 +18,7 @@ type flags struct {
 	signingKey    string
 	signingMethod string
 	configPath    string
-	upstreamURL   string
 	listenAddress string
-	validIssuer   string
-	validAudience string
-	expRequired   string
-	nbfRequired   string
-	clockSkew     string
 }
 
 func main() {
@@ -53,22 +45,6 @@ func main() {
 }
 
 func newServer(f *flags, configReader io.Reader) (*services.Server, error) {
-	var upstream http.Handler
-
-	if f.upstreamURL != "" {
-		// parse upstream URL
-		parsedURL, err := url.Parse(f.upstreamURL)
-		if err != nil {
-			return nil, fmt.Errorf("upstream url could not be parsed: %w", err)
-		}
-
-		if parsedURL.Scheme != "http" && parsedURL.Scheme != "https" {
-			return nil, fmt.Errorf("upstream url scheme must be http or https")
-		}
-
-		upstream = httputil.NewSingleHostReverseProxy(parsedURL)
-	}
-
 	parser := services.YamlConfigParser{}
 	cfg, err := parser.ParseConfig(configReader)
 	if err != nil {
@@ -80,46 +56,34 @@ func newServer(f *flags, configReader io.Reader) (*services.Server, error) {
 		return nil, fmt.Errorf("invalid config: %w", err)
 	}
 
-	var clockSkew int
-	if f.clockSkew != "" {
-		clockSkew, err = strconv.Atoi(f.clockSkew)
-		if err != nil {
-			return nil, fmt.Errorf("clock skew flag %s cannot be converted to integer", f.clockSkew)
-		}
-	} else {
-		clockSkew = 0
+	var upstream http.Handler
+	if cfg.Server.ParsedURL != nil {
+		upstream = httputil.NewSingleHostReverseProxy(cfg.Server.ParsedURL)
 	}
 
 	authenticator, err := services.NewAuthenticator(
 		[]byte(f.signingKey),
 		f.signingMethod,
-		f.validIssuer,
-		f.validAudience,
-		f.expRequired != "false",
-		f.nbfRequired != "false",
-		clockSkew)
+		cfg.Authentication)
 
 	if err != nil {
 		return nil, fmt.Errorf("could not create authenticator: %w", err)
 	}
 
-	s := services.NewServer(
+	return services.NewServer(
 		upstream,
 		services.NewRouteMatcher(cfg.RoutePolicies),
 		services.NewAuthorizer(cfg.ClaimPolicies),
-		authenticator)
-
-	return s, nil
+		authenticator,
+		cfg.Server), nil
 }
 
 func parseFlags() *flags {
 	f := flags{
 		configPath:    "/etc/bouncer/config.yaml",
 		listenAddress: ":3512",
-		expRequired:   "true",
-		nbfRequired:   "true",
 	}
-	
+
 	printVersion := flag.Bool("v", false, "print version and exit")
 	flag.StringVar(&f.signingKey, "k",
 		lookupEnv("BOUNCER_SIGNING_KEY", ""),
@@ -136,30 +100,6 @@ func parseFlags() *flags {
 	flag.StringVar(&f.listenAddress, "l",
 		lookupEnv("BOUNCER_LISTEN_ADDRESS", f.listenAddress),
 		fmt.Sprintf("listen address, default = %s", f.listenAddress))
-
-	flag.StringVar(&f.upstreamURL, "url",
-		lookupEnv("BOUNCER_UPSTREAM_URL", ""),
-		"URL to be called when the request is authorized")
-
-	flag.StringVar(&f.validIssuer, "iss",
-		lookupEnv("BOUNCER_VALID_ISSUER", ""),
-		fmt.Sprintf("valid token issuer"))
-
-	flag.StringVar(&f.validAudience, "aud",
-		lookupEnv("BOUNCER_VALID_AUDIENCE", ""),
-		fmt.Sprintf("valid token audience"))
-
-	flag.StringVar(&f.expRequired, "exp",
-		lookupEnv("BOUNCER_REQUIRE_EXPIRATION", f.expRequired),
-		fmt.Sprintf("require token expiration timestamp claims, default = %s", f.expRequired))
-
-	flag.StringVar(&f.nbfRequired, "nbf",
-		lookupEnv("BOUNCER_REQUIRE_NOT_BEFORE", f.nbfRequired),
-		fmt.Sprintf("require token not before timestamp claims, default = %s", f.nbfRequired))
-
-	flag.StringVar(&f.clockSkew, "clk",
-		lookupEnv("BOUNCER_CLOCK_SKEW", f.clockSkew),
-		fmt.Sprintf("require token not before timestamp claims, default = %s", f.nbfRequired))
 
 	flag.Parse()